feat: Deploy model instances through the Gateway
Allow the user to deploy model instances through the Gateway, by
providing the model URI of a model logged to the tracking server or
the tracking ID corresponding to the run that created it. This is
possible through the new endpoint POST /models/{model_name}, with the
specified model_name being the desired name of the new service. The
Gateway is responsible for spinning up Docker containers running CMS
instances from the provided model. Instances deployed this way are
assigned a TTL (defaults to 1 day), configurable as part of the request.
A separate service, the cmg-ripper, is responsible for cleaning up
expired instances, so that our servers aren't overloaded. If a user
wishes to manage the lifecycle of their model instances themselves, they
can do so by setting the TTL to -1 to skip automatic cleanup.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>