gw: Download results instead of returning a URL
Edit the 'GET /tasks/{task_id}' to allow downloading the object when the
'download' query parameter is set to 'true'. Previously the Gateway
would ask MinIO to generate a presigned URL, which the user could use to
download the object. This change allows the user to download the object
through the Gateway, without the need for direct access to the object
store.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>