tests: Parse and verify MLflow URLs
Add a util for parsing and verifying MLflow URLs. This util will be used
to verify that the URLs found in the results of long-running tasks (e.g.
training) are valid MLflow URLs (i.e. they're reachable URLs and consist
of the expected parts) and extract the tracking URI, experiment ID, and
run ID from them.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>