schd: Requeue task if training server is unavailable
Restructure the logic around server response handling to allow for more
fine-grained control. If the training server is reported unavailable due
to a training job already running, the task is requeued instead of
dropped. A shortcoming of this approach, given that RabbitMQ attempts to
place requeued messages in their original positions or near the head of
the queue, is that long-running tasks (e.g. training/evaluation) can
block the execution of others that could be processed in parallel on the
same servers (CMS only allows running a single training task at a time
to work around resource limitations, but places no restrictions on the
execution of short-running tasks). This could be addressed by
introducing a separate queue for long-running tasks, or by increasing
the priority of short-running tasks in the queue. In any case, we expect
to have a more robust solution once we move to a more sophisticated,
resource and task-aware scheduler.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>