dev: Add script for running the project locally
Introduce 'run_local.py' script for running the project locally. This is
intended for development purposes and should not be used in production.
The script spins up the required external services (e.g. MinIO,
PostgreSQL, RabbitMQ) as Docker containers and starts the Gateway,
Scheduler, and Ripper services as separate local processes. It also
watches the codebase for changes in the Python files, restarting the
services when necessary. Stopping the script (e.g. with Ctrl+C) cleans
up the Docker containers and stops all processes.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>