gw: Add helper for constructing a CMS URL
Add helper to construct a CMS URL given a model name and an optional
endpoint. This function uses the server URL configuration option as
base URL, but defaults to the treating the model name as a Docker
service name if the configuration option is not set, allowing for local
testing without a proxy.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>