api: Download task result if `download` is true
Previously, in order to download the task result via the '/tasks' API,
the client had to set the 'detail' and 'download' query parameters to
true, which was an oversight. This commit fixes this issue by giving
priority to the 'download' parameter when it is set to true. As a
result, the client can now download the task result by only setting
the latter.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>