cmg: Introduce Gateway Python client
Add a new package providing a Python client for the CMG Gateway, which
can be used to perform all operations available through the Gateway API.
The client is designed as asynchronous by default, but a synchronous
wrapper is also provided for convenience. Core functionality is tested
through unit tests, and the package is ready for distribution via PyPI.
In the future, more high-level methods will be added corresponding to
the endpoints exposed by the CMS API (currently we only provide the
most basic and common ones, 'process' and 'redact'); nevertheless, the
generic 'submit_task' method can be used in the meantime to perform any
HTTP request to the Gateway. The client packaged is already available
on PyPI, and can be installed via the following command:
'pip install cogstack-model-gateway-client'

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>