client: Fix GatewayClientSync loop handling issues
Revamp the design of GatewayClientSync to ensure it works in any
environment. If an event loop is already running in the current thread,
it should create a background thread with its own event loop to run the
synchronous code. If no event loop is running, it should create and use
its own in the current thread, taking care of cleaning up afterwards.
Add docstrings and arguments to sync client methods to match their async
counterparts and help with development.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>