chore: Ensure test config.json is ignored in the future
The integration tests update the config.json file when running, adding
the MLflow and MinIO connection details. These are specific to the local
testing environment and likely the given run, and therefore don't need
to be committed to the repository. This commit adds the config.json
file to the .gitignore to prevent accidental commits in the future.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>