feat: Dynamically deploy models to match request targets (#1)
Allow configuring on-demand models through the API, by providing info
necessary for deploying CMS instances automatically when requests are
received targeting a certain model. This includes specifying the model
name, the tracking server artifact URI or tracking ID of the run that
generated it, Docker resource requirements, the time a deployed instance
can stay idle before being removed. Model names must be unique, since
they are used to determine request targets, serving as the Docker
container name for the deployed CMS instances. When a request is
received for a model that does not have a deployed instance, a new
instance is created automatically using the provided configuration- 
if one exists- and the request is routed to it. If no configuration
exists for the requested model, an error is returned. From now on,
available models as listed through the API include both running
instances and on-demand models. This commit also updates the client
library to support creating, updating, deleting, and listing on-demand
configurations, and extends the integration tests to cover the new
endpoints as well as the on-demand deployment functionality.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>